+++ OnlySnarf/lib/driver.py
@@ -1264,25 +1264,13 @@ class Driver:
                 element.send_keys(str(Settings.get_phone_number()))
                 element.send_keys(Keys.ENTER)
 
+            # TODO: requires testing, not successfuly receiving email w/ code to test further
             def try_email():
                 Settings.print("email verification required - please enter the code sent to your email!")
-                # element = self.browser.switch_to.active_element
-                # element.send_keys(str(input("Enter code: ")))
-                # element.send_keys(Keys.SHIFT + Keys.TAB)
-                # element.send_keys(Keys.ENTER)
-
-                elements = self.browser.find_elements(By.TAG_NAME, "a")
-                
-                element = [elem for elem in elements if "I didn\'t receive the email" in str(elem.get_attribute('innerHTML'))][0]
-                element.click()
-
-                # action = ActionChains(self.browser)
-                # action.click(on_element=self.browser.switch_to.active_element)
-                # action.key_down(Keys.SHIFT).send_keys(Keys.TAB).key_up(Keys.SHIFT)
-                # action.send_keys(Keys.ENTER)
-                # action.perform()
-                time.sleep(1)
-                print(self.browser.find_element(By.TAG_NAME, "body").text)
+                element = self.browser.switch_to.active_element
+                element.send_keys(str(input("Enter code: ")))
+                element.send_keys(Keys.SHIFT + Keys.TAB)
+                element.send_keys(Keys.ENTER)
 
             try:
                 Settings.dev_print("waiting for login check...")
@@ -1292,7 +1280,7 @@ class Driver:
                 return True
             except TimeoutException as te:
                 bodyText = self.browser.find_element(By.TAG_NAME, "body").text
-                Settings.dev_print(bodyText)
+                Settings.dev_print()
                 # check for phone number page
                 if "Verify your identity by entering the phone number associated with your Twitter account." in str(bodyText):
                     try_phone()