+++ OnlySnarf/api.py
@@ -16,7 +16,7 @@ def create_app():
             Settings.dev_print(args)
             config["text"] = args["text"]
             config["user"] = args["user"]
-            try: config["files"] = args["input"].split(",")
+            try: config["input"] = args["input"].split(",")
             except Exception as e: pass
             try: config["price"] = args["price"] or 0
             except Exception as e: pass
@@ -24,9 +24,6 @@ def create_app():
             except Exception as e: pass
             try: config["performers"] = args["performers"]
             except Exception as e: pass
-            if app.testing:
-                config["debug"] = True
-                config["verbose"] = 3
             Snarf.message()
             Snarf.close()
         except Exception as e:
@@ -40,7 +37,7 @@ def create_app():
             args = json.loads(request.data)
             Settings.dev_print(args)
             config["text"] = args["text"]
-            try: config["files"] = args["input"].split(",")
+            try: config["input"] = args["input"].split(",")
             except Exception as e: pass
             try: config["performers"] = args["performers"]
             except Exception as e: pass
@@ -52,9 +49,6 @@ def create_app():
             except Exception as e: pass
             try: config["expires"] = args["expires"]
             except Exception as e: pass
-            if app.testing:
-                config["debug"] = True
-                config["verbose"] = 3
             Snarf.post()
             Snarf.close()
         except Exception as e:
@@ -66,8 +60,9 @@ def create_app():
 
 def main():
     app = create_app()
-    app.debug = True
-    app.testing = True
+    if str(config["debug"]) == "True":
+        app.debug = True
+        app.testing = True
     app.run(host="0.0.0.0", port=5000)
 
 if __name__ == "__main__":